Moved ResourceNameField to its own file
